backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
celery==5.3.4
redis==5.0.1
openai==1.54.3
ffmpeg-python==0.2.0
python-multipart==0.0.6
python-dotenv==1.0.0
sqlmodel==0.0.14
httpx==0.27.2
pydantic-settings==2.1.0
psycopg2-binary==2.9.9
alembic==1.13.1
gunicorn==21.2.0